Refactor query param methods
