+++ src/components/SettingsPanel/SettingsPanel.js
@@ -4,8 +4,7 @@ import Transition from 'react-transition-group/Transition';
 import debounce from 'lodash.debounce';
 import { MIN_FONT_SIZE, MAX_FONT_SIZE, COLOR_UPDATE_DELAY } from 'config.js';
 import { getAutoTextColor } from 'utils/colour';
-import { getQueryParams } from 'utils/url';
-import { hasDefaultParams } from 'utils/parse';
+import { getQueryParams, hasDefaultParams } from 'utils/url';
 import { DownArrowIcon, CloseIcon } from 'shared/icons';
 import { Range } from 'components/Range/Range.styles';
 import { ColorPicker } from 'components/ColorPicker/ColorPicker';
