+++ src/utils/parse.test.js
@@ -12,7 +12,6 @@ import {
 	removeTrailingWhitespaceOrBreaks,
 	cleanMarkup,
 	removeDefaultParams,
-	hasDefaultParams,
 } from './parse';
 import {
 	DEFAULT_FONT_SIZE,
@@ -317,32 +316,6 @@ describe('clean markup', () => {
 			const output = cleanMarkup(input);
 			expect(output).toBe(expected);
 		});
-	});
-});
-
-describe('hasDefaultParams', () => {
-	it('handles empty parameters', () => {
-		const input = {};
-		const output = hasDefaultParams(input);
-		expect(output).toBe(true);
-	});
-
-	it('handles params with only text defined', () => {
-		const input = { text: 'test' };
-		const output = hasDefaultParams(input);
-		expect(output).toBe(true);
-	});
-
-	it('handles user-defined params', () => {
-		const input = {
-			text: 'test',
-			bgColor: '#f00',
-			fontColor: '#00f',
-			fontSize: '16',
-			fontStyle: 'monospaced',
-		};
-		const output = hasDefaultParams(input);
-		expect(output).toBe(false);
 	});
 });
