+++ src/utils/url.js
@@ -3,10 +3,7 @@ import pipe from 'ramda/src/pipe';
 import map from 'ramda/src/map';
 import evolve from 'ramda/src/evolve';
 import { fetchShortUrl } from 'api/api';
-import { DEFAULT_PARAMS } from 'config.js';
 import { removeDefaultParams, cleanMarkup } from './parse';
-import { isUndefined } from './type';
-import { getAutoTextColor } from './colour';
 
 export function setUrlParams(newParams) {
 	const params = { ...getQueryParams(), ...newParams };
@@ -22,6 +19,13 @@ export const getQueryParams = () =>
 	qs.parse(document.location.search, {
 		ignoreQueryPrefix: true,
 	});
+
+export function hasDefaultParams(params) {
+	const testParams = { ...params };
+	delete testParams.text;
+
+	return Object.keys(testParams).length === 0;
+}
 
 const cleanParams = pipe(
 	evolve({ text: cleanMarkup }),
@@ -45,19 +49,4 @@ export function getShareUrl(urlToShare) {
 			return data.url;
 		})
 		.catch(() => urlToShare);
-}
-
-export function deriveStateFromQueryParams() {
-	const queryParams = getQueryParams();
-	const isAutoFontColor = isUndefined(queryParams.fontColor);
-
-	if (isAutoFontColor) {
-		queryParams.fontColor = getAutoTextColor(queryParams.bgColor);
-	}
-
-	return {
-		...DEFAULT_PARAMS,
-		...queryParams,
-		isAutoFontColor,
-	};
 }
